Refactor into Module
* created 2 directories (db, utils)
* moved utility functions out of nbmon and into util.py
